Split api/index.ts into components
