+++ web/src/api/meta.ts
@@ -1,0 +1,4 @@
+export interface ObjectMeta {
+  namespace: string;
+  name: string;
+}
