+++ web/src/api/mock-utils.ts
@@ -1,0 +1,3 @@
+export function delay(amount: number): Promise<void> {
+  return new Promise((resolve) => setTimeout(resolve, amount));
+}
